Update 3_Generate Input.py
Update of the following functions: 
- loc_file_RMS : fillna(0) for CVXVAL columns
- loc_file_AIR : fillna(0) for CVXVAL columns

Previously, there was an error between respectively the sum of BDLG / CONT and the sum of EQCV1VAL and EQCV2VAL. 

The issue came from the value of EQCV3val that show for some rows None value and the filter on positive values remove these rows. By replacing the None values by 0, these rows are now kept, and the calculation of the rows are now exact.